# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.DiagnosisCategory:
    properties:
      clinic_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      names:
        description: Relations
        items:
          $ref: '#/definitions/model.DiagnosisName'
        type: array
      sort_order:
        type: integer
      updated_at:
        type: string
    type: object
  model.DiagnosisName:
    properties:
      category:
        allOf:
        - $ref: '#/definitions/model.DiagnosisCategory'
        description: Relations
      clinic_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      diagnosis_category_id:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
    type: object
...
  model.ExamType:
    properties:
      clinic_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      items:
        description: Relations
        items:
          $ref: '#/definitions/model.ExamTypeItem'
        type: array
      name:
        type: string
      price:
        type: number
      sort_order:
        type: integer
      updated_at:
        type: string
    type: object
  model.ExamTypeItem:
    properties:
      created_at:
        type: string
      exam_type_id:
        type: integer
      id:
        type: integer
      inspection_value:
        type: string
      name:
        type: string
      normal_value:
        type: string
      sort_order:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 動物病院 電子カルテシステム API（45テーブル・専用マスタテーブル版）
  title: Animal Ekarte API
  version: "2.0"
paths:
...
  /masters/diagnosis-categories:
    get:
      consumes:
      - application/json
      description: 登録されている診断カテゴリの一覧を返す
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DiagnosisCategory'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 診断カテゴリ一覧取得
      tags:
      - Masters
    post:
      consumes:
      - application/json
      description: 新しい診断カテゴリを作成する
      parameters:
      - description: 診断カテゴリ情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.DiagnosisCategory'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.DiagnosisCategory'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 診断カテゴリ作成
      tags:
      - Masters
  /masters/diagnosis-categories/{id}:
    delete:
      consumes:
      - application/json
      description: 指定IDの診断カテゴリを削除する
      parameters:
      - description: 診断カテゴリID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 診断カテゴリ削除
      tags:
      - Masters
    put:
      consumes:
      - application/json
      description: 指定IDの診断カテゴリを更新する
      parameters:
      - description: 診断カテゴリID
        in: path
        name: id
        required: true
        type: integer
      - description: 診断カテゴリ情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.DiagnosisCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DiagnosisCategory'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 診断カテゴリ更新
      tags:
      - Masters
  /masters/diagnosis-names:
    get:
      consumes:
      - application/json
      description: 登録されている診断名の一覧を返す。category_idクエリパラメータでカテゴリ絞り込み可能
      parameters:
      - description: 診断カテゴリID（省略時は全件返す）
        in: query
        name: category_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DiagnosisName'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 診断名一覧取得
      tags:
      - Masters
    post:
      consumes:
      - application/json
      description: 新しい診断名を作成する
      parameters:
      - description: 診断名情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.DiagnosisName'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.DiagnosisName'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 診断名作成
      tags:
      - Masters
  /masters/diagnosis-names/{id}:
    delete:
      consumes:
      - application/json
      description: 指定IDの診断名を削除する
      parameters:
      - description: 診断名ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 診断名削除
      tags:
      - Masters
    put:
      consumes:
      - application/json
      description: 指定IDの診断名を更新する
      parameters:
      - description: 診断名ID
        in: path
        name: id
        required: true
        type: integer
      - description: 診断名情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.DiagnosisName'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DiagnosisName'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 診断名更新
      tags:
      - Masters
  /masters/examination-types:
    get:
      consumes:
      - application/json
      description: 登録されている検査種別の一覧を返す
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ExamType'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 検査種別一覧取得
      tags:
      - Masters
    post:
      consumes:
      - application/json
      description: 新しい検査種別を作成する
      parameters:
      - description: 検査種別情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ExamType'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ExamType'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 検査種別作成
      tags:
      - Masters
  /masters/examination-types/{id}:
    delete:
      consumes:
      - application/json
      description: 指定IDの検査種別を削除する
      parameters:
      - description: 検査種別ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 検査種別削除
      tags:
      - Masters
    put:
      consumes:
      - application/json
      description: 指定IDの検査種別を更新する
      parameters:
      - description: 検査種別ID
        in: path
        name: id
        required: true
        type: integer
      - description: 検査種別情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ExamType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ExamType'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 検査種別更新
      tags:
      - Masters
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"